+++ src/models/Filter.js
@@ -126,4 +126,62 @@ export default class Filter {
       return true;
     }
   };
+
+  /*
+   * A fn to count the number of current applied filters
+  */
+  countAppliedFilters() {
+    let numFilters = 0;
+
+    if (this.active !== undefined && this.approved !== undefined) {
+      numFilters += 1;
+    }
+
+    if (this.uuid) {
+      numFilters += 1;
+    }
+
+    if (this.captureId) {
+      numFilters += 1;
+    }
+
+    if (this.deviceIdentifier) {
+      numFilters += 1;
+    }
+
+    if (this.planterId) {
+      numFilters += 1;
+    }
+
+    if (this.planterIdentifier) {
+      numFilters += 1;
+    }
+
+    if (this.tagId > 0) {
+      numFilters += 1;
+    }
+
+    if (this.dateStart) {
+      numFilters += 1;
+    }
+
+    if (this.dateEnd) {
+      numFilters += 1;
+    }
+
+    if (this.organizationId && this.organizationId !== ALL_ORGANIZATIONS) {
+      numFilters += 1;
+    }
+
+    if (this.speciesId && this.speciesId !== ALL_SPECIES) {
+      numFilters += 1;
+    }
+
+    if (this.tokenId && this.tokenId !== 'All') {
+      numFilters += 1;
+    }
+
+    return numFilters;
+  }
+
 }